+++ src/main/java/com/github/resalner/javapractice/service/OrderItemServiceImpl.java
@@ -31,42 +31,41 @@ public class OrderItemServiceImpl implements OrderItemService {
 
 	@Override
 	public OrderItem getOrderItem(long id) {
-		return orderItemRepository.findById(id)
-				.orElseThrow(() -> new EntityNotFoundException("не найден элемент заказа с id = " + id));
+		return getOrderItemIfExists(id);
 	}
 
 	@Override
 	public void deleteOrderItem(long id) {
+		if (!orderItemRepository.existsById(id)) {
+			throw new EntityNotFoundException("Не найден продукт с id = " + id);
+		}
 		orderItemRepository.deleteById(id);
 	}
 
 	@Override
 	public OrderItem updateOrderItem(long id, OrderItem orderItemForUpdate) {
-		OrderItem orderItem = orderItemRepository.findById(id)
-				.orElseThrow(() -> new EntityNotFoundException("не найден элемент заказа с id = " + id));
+		OrderItem orderItem = getOrderItemIfExists(id);
 
 		Order newOrder = orderItemForUpdate.getOrder();
 		Product newProduct = orderItemForUpdate.getProduct();
 		Integer newCount = orderItemForUpdate.getCount();
 		Double newPrice = orderItemForUpdate.getPrice();
 
-		if (Objects.nonNull(newOrder)) {
+		orderItem.setOrder(newOrder);
+		orderItem.setProduct(newProduct);
+		orderItem.setCount(newCount);
+		orderItem.setPrice(newPrice);
 
-			orderItem.setOrder(newOrder);
-		}
-		if (Objects.nonNull(newProduct)) {
-
-			orderItem.setProduct(newProduct);
-		}
-		if (Objects.nonNull(newCount)) {
-
-			orderItem.setCount(newCount);
-		}
-		if (Objects.nonNull(newPrice)) {
-
-			orderItem.setPrice(newPrice);
-		}
 		orderItem = orderItemRepository.save(orderItem);
+		
+		return orderItem;
+	}
+	
+	private OrderItem getOrderItemIfExists(long id)
+	{
+		OrderItem orderItem = orderItemRepository.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException("не найден элемент заказа с id = " + id));
+		
 		return orderItem;
 	}
 }
