+++ src/main/java/com/github/resalner/javapractice/service/UserInfoServiceImpl.java
@@ -31,19 +31,20 @@ public class UserInfoServiceImpl implements UserInfoService {
 
 	@Override
 	public UserInfo getUserInfo(long id) {
-		return userInfoRepository.findById(id)
-				.orElseThrow(() -> new EntityNotFoundException("не найден пользователь с id = " + id));
+		return getUserInfoIfExists(id);
 	}
 
 	@Override
 	public void deleteUserInfo(long id) {
+		if (!userInfoRepository.existsById(id)) {
+			throw new EntityNotFoundException("Не найден продукт с id = " + id);
+		}
 		userInfoRepository.deleteById(id);
 	}
 
 	@Override
 	public UserInfo updateUserInfo(long id, UserInfo userInfoForUpdate) {
-		UserInfo userInfo = userInfoRepository.findById(id)
-				.orElseThrow(() -> new EntityNotFoundException("не найден пользователь с id = " + id));
+		UserInfo userInfo = getUserInfoIfExists(id);
 
 		String name = userInfoForUpdate.getName();
 		String surname = userInfoForUpdate.getSurname();
@@ -52,31 +53,22 @@ public class UserInfoServiceImpl implements UserInfoService {
 		boolean gender = userInfoForUpdate.isGender();
 		String email = userInfoForUpdate.getEmail();
 
-		if (Objects.nonNull(name) && !"".equals(name)) {
-
-			userInfo.setName(name);
-		}
-		if (Objects.nonNull(surname) && !"".equals(surname)) {
-
-			userInfo.setSurname(surname);
-		}
-		if (Objects.nonNull(phoneNumber) && !"".equals(phoneNumber)) {
-
-			userInfo.setPhoneNumber(phoneNumber);
-		}
-		if (Objects.nonNull(birthDate)) {
-
-			userInfo.setBirthDate(birthDate);
-		}
-		if (Objects.nonNull(gender)) {
-
-			userInfo.setGender(gender);
-		}
-		if (Objects.nonNull(email) && !"".equals(email)) {
-
-			userInfo.setEmail(email);
-		}
+		userInfo.setName(name);
+		userInfo.setSurname(surname);
+		userInfo.setPhoneNumber(phoneNumber);
+		userInfo.setBirthDate(birthDate);
+		userInfo.setGender(gender);
+		userInfo.setEmail(email);
+		
 		userInfo = userInfoRepository.save(userInfo);
+		
+		return userInfo;
+	}
+
+	private UserInfo getUserInfoIfExists(long id) {
+		UserInfo userInfo = userInfoRepository.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException("не найден пользователь с id = " + id));
+
 		return userInfo;
 	}
 }
